Move utility/setup functions in frontend to separate modules
The modules currently have the following responsibilities:
- `__init__` - only imports the other modules (including plugins) to
  ensure that all routes, filters etc. have been registered
- `base` - initialization of the app and db, definition of the session
  and basic globals
- `filters` - jinja2 filters for templates
- `model_additions` - properties monkey-patched on top of models, such
  as state icons
- `template_functions` - utility functions primarily aimed to be used
  within templates
- `util` - utility functions for flash messages and ordering
- `tabs` - definition of tab menu
- `forms` - form definitions and utilities
- `auth` - authentication
- `api` - API
- `views` - the actual controller part, could be further split
