+++ test/web_test.py
@@ -18,15 +18,8 @@
 
 import re
 
-# pylint:disable = unused-import
-import koschei.frontend.api
-import koschei.frontend.views
-import koschei.frontend.auth
-
-from koschei.frontend import app, db
-from koschei.models import User, PackageGroup, AppliedChange
-
-from test.common import DBTest, my_vcr
+from koschei.models import PackageGroup
+from test.common import my_vcr
 from test.frontend_common import FrontendTest, authenticate, authenticate_admin
